a_cordes/requirements.txt
amqp==5.1.1
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
black==22.3.0
boto==2.49.0
boto3==1.20.21
botocore==1.23.21
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
Django==3.2.11
django-cors-headers==3.7.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
iniconfig==1.1.1
jmespath==0.10.0
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2021.9.24
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
